requirements.txt
fastapi==0.63.0
gunicorn==20.0.4
httpx==0.16.1
uvicorn==0.13.3
pydantic==1.7.3
databases[mysql]==0.4.1
uvloop==0.15.2 
httptools==0.1.1
python-dotenv==0.15.0
requests==2.25.1
PyPika==0.48.9
pymysql==0.10.1
pytest==6.2.2
pytest-cov==2.11.1
validators==0.19.0
email-validator==1.3.1
cryptography==39.0.1
PyJWT==2.6.0
pymemcache==4.0.0
